getRAPIDModulesInfo() and getRAPIDTasks() moved to v1_0::rw::RAPIDService.
