+++ CMakeLists.txt
@@ -48,7 +48,8 @@ set(
     src/rws.cpp
     src/parsing.cpp
     src/connection_options.cpp
-    src/rapid_execution_state.cpp
+
+    src/common/rw/rapid.cpp
 
     src/v1_0/rws.cpp
     src/v1_0/rws_client.cpp
